refactor(frontend): move decodeQrCode util to lib
