+++ src/frontend/src/icp/components/send/IcSendTokenWizard.svelte
@@ -20,7 +20,6 @@
 	} from '$icp/stores/ethereum-fee.store';
 	import type { IcToken } from '$icp/types/ic';
 	import type { IcTransferParams } from '$icp/types/ic-send';
-	import { icDecodeQrCode } from '$icp/utils/qr-code.utils';
 	import { SEND_CONTEXT_KEY, type SendContext } from '$icp-eth/stores/send.store';
 	import {
 		isConvertCkErc20ToErc20,
@@ -58,6 +57,7 @@
 	import { invalidAmount, isNullishOrEmpty } from '$lib/utils/input.utils';
 	import { isNetworkIdBitcoin } from '$lib/utils/network.utils';
 	import { parseToken } from '$lib/utils/parse.utils';
+	import { decodeQrCode } from '$lib/utils/qr-code.utils';
 
 	/**
 	 * Props
@@ -238,7 +238,7 @@
 				expectedToken={$sendToken}
 				bind:destination
 				bind:amount
-				decodeQrCode={icDecodeQrCode}
+				{decodeQrCode}
 				on:icQRCodeBack
 			/>
 		{:else}
