+++ src/frontend/src/lib/utils/qr-code.utils.ts
@@ -2,8 +2,12 @@ import {
 	DecodedUrnSchema,
 	URN_NUMERIC_PARAMS,
 	URN_STRING_PARAMS,
-	type DecodedUrn
+	type DecodedUrn,
+	type QrResponse,
+	type QrStatus
 } from '$lib/types/qr-code';
+import type { OptionToken } from '$lib/types/token';
+import { decodePayment } from '@dfinity/ledger-icrc';
 import { isNullish, nonNullish } from '@dfinity/utils';
 
 /**
@@ -80,4 +84,36 @@ export const decodeQrCodeUrn = (urn: string): DecodedUrn | undefined => {
 		return undefined;
 	}
 	return result.data;
+};
+
+export const decodeQrCode = ({
+	status,
+	code,
+	expectedToken
+}: {
+	status: QrStatus;
+	code?: string | undefined;
+	expectedToken?: OptionToken;
+}): QrResponse => {
+	if (status !== 'success') {
+		return { status };
+	}
+
+	if (isNullish(code)) {
+		return { status: 'cancelled' };
+	}
+
+	const payment = decodePayment(code);
+
+	if (isNullish(payment)) {
+		return { status: 'success', destination: code };
+	}
+
+	const { token: symbol, identifier: destination, amount } = payment;
+
+	if (nonNullish(expectedToken) && symbol.toLowerCase() !== expectedToken.symbol.toLowerCase()) {
+		return { status: 'token_incompatible' };
+	}
+
+	return { status: 'success', destination, symbol, amount };
 };
