+++ src/frontend/src/tests/lib/utils/qr-code.utils.spec.ts
@@ -1,10 +1,17 @@
+import { ICP_TOKEN } from '$env/tokens.env';
 import type { EthereumNetwork } from '$eth/types/network';
 import { tokens } from '$lib/derived/tokens.derived';
 import type { DecodedUrn } from '$lib/types/qr-code';
-import { decodeQrCodeUrn } from '$lib/utils/qr-code.utils';
+import { decodeQrCode, decodeQrCodeUrn } from '$lib/utils/qr-code.utils';
 import { generateUrn } from '$tests/mocks/qr-generator.mock';
+import { decodePayment } from '@dfinity/ledger-icrc';
 import { assertNonNullish } from '@dfinity/utils';
 import { get } from 'svelte/store';
+import type { MockedFunction } from 'vitest';
+
+vi.mock('@dfinity/ledger-icrc', () => ({
+	decodePayment: vi.fn()
+}));
 
 describe('decodeUrn', () => {
 	const tokenList = get(tokens);
@@ -47,5 +54,67 @@ describe('decodeUrn', () => {
 
 			expect(result).toEqual(expectedResult);
 		});
+	});
+});
+
+describe('decodeQrCode', () => {
+	const token = ICP_TOKEN;
+	const address = 'some-address';
+	const amount = 1.23;
+	const code = 'some-qr-code';
+	const mockDecodePayment = decodePayment as MockedFunction<typeof decodePayment>;
+
+	beforeEach(() => {
+		vi.clearAllMocks();
+	});
+
+	it('should return { result } when result is not success', () => {
+		const response = decodeQrCode({ status: 'cancelled' });
+		expect(response).toEqual({ status: 'cancelled' });
+	});
+
+	it('should return { status: "cancelled" } when code is nullish', () => {
+		const response = decodeQrCode({ status: 'success', code: undefined });
+		expect(response).toEqual({ status: 'cancelled' });
+	});
+
+	it('should return { status: "success", identifier: code } when payment is nullish', () => {
+		mockDecodePayment.mockReturnValue(undefined);
+
+		const response = decodeQrCode({ status: 'success', code });
+		expect(response).toEqual({ status: 'success', destination: code });
+
+		expect(mockDecodePayment).toHaveBeenCalledWith(code);
+	});
+
+	it('should return { status: "token_incompatible" } when tokens do not match', () => {
+		const payment = {
+			token: `not-${token.symbol}`,
+			identifier: address,
+			amount: amount
+		};
+		mockDecodePayment.mockReturnValue(payment);
+
+		const response = decodeQrCode({ status: 'success', code: code, expectedToken: token });
+		expect(response).toEqual({ status: 'token_incompatible' });
+		expect(mockDecodePayment).toHaveBeenCalledWith(code);
+	});
+
+	it('should return { status: "success", identifier, token, amount } when everything matches', () => {
+		const payment = {
+			token: token.symbol,
+			identifier: address,
+			amount: amount
+		};
+		mockDecodePayment.mockReturnValue(payment);
+
+		const response = decodeQrCode({ status: 'success', code: code, expectedToken: token });
+		expect(response).toEqual({
+			status: 'success',
+			destination: address,
+			symbol: token.symbol,
+			amount: amount
+		});
+		expect(mockDecodePayment).toHaveBeenCalledWith(code);
 	});
 });
